Add GraphXJobTest trait
